+++ collect_data/utils/common/datapipeline.py
@@ -19,19 +19,18 @@ from sqlalchemy.schema import MetaData
 from sqlalchemy.dialects.mysql import insert as mysql_insert
 from sqlalchemy.sql.expression import func as sql_func
 
-'''
-'''
-class DataPipeline:
 
-    def __init__(self,config,drivername='mysql+pymysql'):
+class DataPipeline:
+    '''
+    '''
+    def __init__(self, config, drivername='mysql+pymysql', write_google=True):
         '''
         '''
         self.drive_params = config["GoogleDrive"]
-        
-        params = config["parameters"]
-        required_fields = ["table_name","input_db","output_db",
-                           "src","description","github"]
-        
+        self.write_google = write_google
+        # params = config["parameters"]
+        required_fields = ["table_name", "input_db", "output_db",
+                           "src", "description", "github"]
         # Collect the required fields together
         self.metainfo = {}
         for field in required_fields:
@@ -40,32 +39,33 @@ class DataPipeline:
             elif field in config["DEFAULT"]:
                 self.metainfo[field] = config["DEFAULT"][field]
             else:
-                raise RuntimeError("Could not find field "+field+" in configuration.")
-        
+                raise RuntimeError("Could not find field " + field +
+                                   " in configuration.")
         # Connect to the database
         output_db = self.metainfo["output_db"]
         db_cnf = URL(drivername=drivername,
-                     query={'read_default_file':config["DEFAULT"][output_db]})
+                     query={'read_default_file': config["DEFAULT"][output_db]})
         self.engine = create_engine(name_or_url=db_cnf)
         self.conn = self.engine.connect()
         self.metadata = MetaData(bind=self.engine)
-        
+
     def __enter__(self):
         '''Dummy method for with'''
         return self
 
     def __exit__(self, _type, _value, _traceback):
-        """ 
-        """        
+        """
+        """
         # Don't write metadata if there's an error
         if _type is None:
             self.update_metadata_db()
-            # Write the main table and metadata to Drive
-            for table_name in [self.metainfo["table_name"],"metadata"]:
-                file_path = write_json(self.conn,table_name,destination="/tmp/")
-                save_to_drive(file_path,self.drive_params)
-                os.remove(file_path)
-                
+            if self.write_google:
+                # Write the main table and metadata to Drive
+                for table_name in [self.metainfo["table_name"], "metadata"]:
+                    file_path = write_json(self.conn, table_name,
+                                           destination="/tmp/")
+                    save_to_drive(file_path, self.drive_params)
+                    os.remove(file_path)
         # Close the engine connection
         self.conn.close()
 
@@ -74,29 +74,25 @@ class DataPipeline:
         '''
         # Otherwise generate the metadata
         metadata = MetaData(bind=self.engine)
-        soft_table = Table(sql_text('metadata'),metadata,
-                           autoload=True,mysql_charset='utf8',)        
-        
+        soft_table = Table(sql_text('metadata'), metadata,
+                           autoload=True, mysql_charset='utf8',)
         # Get the source(s), to be looped over
         sources = self.metainfo.pop("src")
         if type(sources) is not list:
             sources = [sources]
-
-        # Insert each row of metadata (one per source)        
+        # Insert each row of metadata (one per source)
         logging.info("\tWriting to table metadata")
         self.metainfo["timestamp_update"] = sql_func.current_timestamp()
         for src in sources:
             self.metainfo["src"] = src
-            _insert = mysql_insert(soft_table,values=self.metainfo)            
+            _insert = mysql_insert(soft_table, values=self.metainfo)
             _insert = _insert.on_duplicate_key_update(**self.metainfo)
             self.conn.execute(_insert)
 
-    def insert(self,values):
+    def insert(self, values):
         '''
         '''
-        table = Table(sql_text(self.metainfo["table_name"]),self.metadata,
-                     autoload=True,mysql_charset='utf8',)
+        table = Table(sql_text(self.metainfo["table_name"]), self.metadata,
+                      autoload=True, mysql_charset='utf8',)
         _insert_stmt = sql_insert(table).prefix_with("IGNORE")
         self.conn.execute(_insert_stmt.values(**values))
-        
-
